+++ derive_builder_core/src/build_method.rs
@@ -58,6 +58,9 @@ pub struct BuildMethod<'a> {
     ///
     /// This will be in scope for all initializers as `__default`.
     pub default_struct: Option<Block>,
+    /// Validation function with signature `&FooBuilder -> Result<(), String>`
+    /// to call before the macro-provided struct buildout.
+    pub validation_fn: Option<&'a syn::Path>,
 }
 
 impl<'a> ToTokens for BuildMethod<'a> {
@@ -79,6 +82,9 @@ impl<'a> ToTokens for BuildMethod<'a> {
                      let ident = syn::Ident::new(DEFAULT_STRUCT_NAME);
                      quote!(let #ident: #target_ty = #default_expr;)
                  });
+        let validation_fn = self.validation_fn
+            .as_ref()
+            .map(|vfn| quote!(#vfn(&self)?;));
         let result = self.bindings.result_ty();
         let string = self.bindings.string_ty();
 
@@ -89,6 +95,7 @@ impl<'a> ToTokens for BuildMethod<'a> {
                 #vis fn #ident(#self_param)
                     -> #result<#target_ty #target_ty_generics, #string>
                 {
+                    #validation_fn
                     #default_struct
                     Ok(#target_ty {
                         #(#initializers)*
@@ -134,6 +141,7 @@ macro_rules! default_build_method {
             doc_comment: None,
             bindings: Default::default(),
             default_struct: None,
+            validation_fn: None,
         }
     }
 }
@@ -184,7 +192,7 @@ mod tests {
             }
         ));
     }
-    
+
     #[test]
     fn skip() {
         let mut build_method = default_build_method!();
@@ -192,13 +200,13 @@ mod tests {
         
         assert_eq!(quote!(#build_method), quote!());
     }
-    
+
     #[test]
     fn rename() {
         let ident = syn::Ident::new("finish");
         let mut build_method : BuildMethod = default_build_method!();
         build_method.ident = &ident;
-        
+
         assert_eq!(quote!(#build_method), quote!(
             pub fn finish(&self) -> ::std::result::Result<Foo, ::std::string::String> {
                 Ok(Foo {
@@ -206,5 +214,24 @@ mod tests {
                 })
             }
         ))
+    }
+
+    #[test]
+    fn validation() {
+        let validator_path = syn::parse_path("IpsumBuilder::validate")
+            .expect("Statically-entered path should be valid");
+            
+        let mut build_method: BuildMethod = default_build_method!();
+        build_method.validation_fn = Some(&validator_path);
+
+        assert_eq!(quote!(#build_method), quote!(
+            pub fn build(&self) -> ::std::result::Result<Foo, ::std::string::String> {
+                IpsumBuilder::validate(&self)?;
+
+                Ok(Foo {
+                    foo: self.foo,
+                })
+            }
+        ));
     }
 }